[web-runtime] change base url to src and reorganize shape files
