+++ react-renderer/src/componentMap.tsx
@@ -1,24 +1,24 @@
 // Map between "tag" and corresponding component
-import Circle from "./Circle";
-import Image from "./Image";
-import Label from "./Label";
-import LabelTransform from "./LabelTransform";
-import Rectangle from "./Rectangle";
-import Arc from "./Arc";
-import RectangleTransform from "./RectangleTransform";
-import CircleTransform from "./CircleTransform";
-import Ellipse from "./Ellipse";
-import EllipseTransform from "./EllipseTransform";
-import Square from "./Square";
-import SquareTransform from "./SquareTransform";
-import Curve from "./Curve";
-import CurveTransform from "./CurveTransform";
-import Arrow from "./Arrow";
-import Line from "./Line";
-import LineTransform from "./LineTransform";
-import Polygon from "./Polygon";
-import ImageTransform from "./ImageTransform";
-import ParallelogramTransform from "./ParallelogramTransform";
+import Circle from "shapes/Circle";
+import Image from "shapes/Image";
+import Label from "shapes/Label";
+import LabelTransform from "shapes/LabelTransform";
+import Rectangle from "shapes/Rectangle";
+import Arc from "shapes/Arc";
+import RectangleTransform from "shapes/RectangleTransform";
+import CircleTransform from "shapes/CircleTransform";
+import Ellipse from "shapes/Ellipse";
+import EllipseTransform from "shapes/EllipseTransform";
+import Square from "shapes/Square";
+import SquareTransform from "shapes/SquareTransform";
+import Curve from "shapes/Curve";
+import CurveTransform from "shapes/CurveTransform";
+import Arrow from "shapes/Arrow";
+import Line from "shapes/Line";
+import LineTransform from "shapes/LineTransform";
+import Polygon from "shapes/Polygon";
+import ImageTransform from "shapes/ImageTransform";
+import ParallelogramTransform from "shapes/ParallelogramTransform";
 import draggable from "./Draggable";
 
 // prettier-ignore
