+++ react-renderer/src/inspector/views/ShapeView.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import IViewProps from "./IViewProps";
-import { staticMap } from "src/componentMap";
+import { staticMap } from "componentMap";
 import styled from "styled-components";
 import { ObjectInspector } from "react-inspector";
 
@@ -44,9 +44,7 @@ class ShapeView extends React.Component<IViewProps, IState> {
           overflow: "hidden",
         }}
       >
-        <div
-          style={{ overflowY: "auto", height: "100%", }}
-        >
+        <div style={{ overflowY: "auto", height: "100%" }}>
           <ul
             style={{
               listStyleType: "none",
@@ -59,41 +57,47 @@ class ShapeView extends React.Component<IViewProps, IState> {
           >
             {frame.shapes.map(
               ({ properties, shapeType }: Shape, key: number) => {
-		  // If the inspector is crashing around here, then probably the shape doesn't have the width/height properties, so add a special case as below
-		  // console.log("properties, shapeType", properties, shapeType, properties.w, properties.h);
+                // If the inspector is crashing around here, then probably the shape doesn't have the width/height properties, so add a special case as below
+                // console.log("properties, shapeType", properties, shapeType, properties.w, properties.h);
 
-                  let [w, h] = [0, 0];
+                let [w, h] = [0, 0];
 
-                  if (shapeType === "Circle") {
-		      [w, h] = 
-			  [
-                              (properties.r.contents as number) * 2,
-                              (properties.r.contents as number) * 2,
-			  ];
-		  } else if (shapeType === "Square") {
-		      [w, h] = 
-			  [
-			      (properties.side.contents as number),
-			      (properties.side.contents as number),
-			  ];
-		  } else if (shapeType === "Arrow") {
-		      const [sx, sy, ex, ey] = [properties.startX.contents as number, 
-						properties.startY.contents as number, 
-						properties.endX.contents as number, 
-						properties.endY.contents as number];
+                if (shapeType === "Circle") {
+                  [w, h] = [
+                    (properties.r.contents as number) * 2,
+                    (properties.r.contents as number) * 2,
+                  ];
+                } else if (shapeType === "Square") {
+                  [w, h] = [
+                    properties.side.contents as number,
+                    properties.side.contents as number,
+                  ];
+                } else if (shapeType === "Arrow") {
+                  const [sx, sy, ex, ey] = [
+                    properties.startX.contents as number,
+                    properties.startY.contents as number,
+                    properties.endX.contents as number,
+                    properties.endY.contents as number,
+                  ];
 
-		      const padding = 50; // Because arrow may be horizontal or vertical, and we don't want the size to be zero in that case
-		      // size of bbox of arrow
-		      [w, h] = [Math.max(Math.abs(ex - sx), padding), Math.max(Math.abs(ey - sy), padding)];
-		  } else if (shapeType === "Curve") {
-                      console.log("TODO: Curve bbox");
-		      // size of bbox of arrow
-		      [w, h] = [100, 100];
-		  } else {
-		      [w, h] = [properties.w.contents as number, properties.h.contents as number];
-		  }
+                  const padding = 50; // Because arrow may be horizontal or vertical, and we don't want the size to be zero in that case
+                  // size of bbox of arrow
+                  [w, h] = [
+                    Math.max(Math.abs(ex - sx), padding),
+                    Math.max(Math.abs(ey - sy), padding),
+                  ];
+                } else if (shapeType === "Curve") {
+                  console.log("TODO: Curve bbox");
+                  // size of bbox of arrow
+                  [w, h] = [100, 100];
+                } else {
+                  [w, h] = [
+                    properties.w.contents as number,
+                    properties.h.contents as number,
+                  ];
+                }
 
-                  return (
+                return (
                   <ShapeItem
                     key={`shapePreview-${key}`}
                     selected={selectedShape === key}
@@ -127,7 +131,7 @@ class ShapeView extends React.Component<IViewProps, IState> {
             overflow: "auto",
             height: "100%",
             flexGrow: 1,
-            boxSizing: "border-box"
+            boxSizing: "border-box",
           }}
         >
           {frame.shapes[selectedShape] && (
