+++ react-renderer/src/shapes/CircleTransform.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { toHex, svgTransformString } from "./Util";
-import { IGPIProps } from "./types";
+import { toHex, svgTransformString } from "utils/Util";
+import { IGPIProps } from "types";
 
-class RectangleTransform extends React.Component<IGPIProps> {
+class CircleTransform extends React.Component<IGPIProps> {
   public render() {
     const { shape } = this.props;
     const { canvasSize } = this.props;
@@ -18,14 +18,11 @@ class RectangleTransform extends React.Component<IGPIProps> {
       canvasSize
     );
 
-    // The default rectangle is an axis-aligned unit square centered at the origin
-    // Its position, size, angle, etc. is all set by the Penrose transform
     return (
-      <rect
-        x={-0.5}
-        y={-0.5}
-        width={1.0}
-        height={1.0}
+      <circle
+        cx={-0.0}
+        cy={-0.0}
+        r={1.0}
         fill={fillColor}
         fillOpacity={fillAlpha}
         stroke={strokeColor}
@@ -35,8 +32,9 @@ class RectangleTransform extends React.Component<IGPIProps> {
         transform={transformStr}
       >
         <title>{shape.name.contents}</title>
-      </rect>
+        <desc>Circle representing {shape.name.contents}</desc>
+      </circle>
     );
   }
 }
-export default RectangleTransform;
+export default CircleTransform;
