+++ react-renderer/src/shapes/Curve.tsx
@@ -1,12 +1,12 @@
 import * as React from "react";
-import { svgTransformString, toHex, Arrowhead } from "./Util";
+import { toScreen, toHex, Arrowhead, Shadow } from "utils/Util";
 import { flatten } from "lodash";
-import { IGPIProps } from "./types";
+import { IGPIProps } from "types";
 
 const toCmdString = (cmd: any, canvasSize: [number, number]) => {
   switch (cmd.tag) {
     case "Pt":
-      return "L" + cmd.contents.join(" ");
+      return "L" + toScreen(cmd.contents, canvasSize).join(" ");
     case "CubicBez":
       return pathCommandString("C", cmd.contents, canvasSize);
     case "CubicBezJoin":
@@ -28,19 +28,25 @@ const pathCommandString = (
   command +
   flatten(
     pts.map((coords: [number, number]) => {
-      return coords; // TODO factor this out
+      return toScreen(coords, canvasSize);
     })
   ).join(" ");
 
 const fstCmdString = (pathCmd: any, canvasSize: [number, number]) => {
   if (pathCmd.tag === "Pt") {
-    return "M" + pathCmd.contents.join(" ");
+    return "M" + toScreen(pathCmd.contents, canvasSize).join(" ");
   } else {
     return toCmdString(pathCmd, canvasSize);
   }
 };
 
 const toSubPathString = (commands: any[], canvasSize: [number, number]) => {
+  // TODO: deal with an empty list more gracefully. This next line will crash with undefined head command if empty.
+  if (!commands || !commands.length) {
+    console.error("WARNING: empty path");
+    return "";
+  }
+
   const [headCommand, ...tailCommands] = commands;
   return (
     fstCmdString(headCommand, canvasSize) +
@@ -57,49 +63,43 @@ const toPathString = (pathData: any[], canvasSize: [number, number]) =>
     })
     .join(" ");
 
-class CurveTransform extends React.Component<IGPIProps> {
+class Curve extends React.Component<IGPIProps> {
   public render() {
     const { shape } = this.props;
     const { canvasSize } = this.props;
     const strokeWidth = shape.strokeWidth.contents;
     const strokeColor = toHex(shape.color.contents);
-    console.log("strokeColor", shape.color.contents, strokeColor);
     const fillColor = toHex(shape.fill.contents);
     const strokeOpacity = shape.color.contents.contents[3];
-    const fillOpacity = shape.fill.contents[3];
+    const fillOpacity = shape.fill.contents.contents[3];
     const arrowheadStyle = shape.arrowheadStyle.contents;
     const arrowheadSize = shape.arrowheadSize.contents;
 
     const leftArrowId = shape.name.contents + "-leftArrowhead";
     const rightArrowId = shape.name.contents + "-rightArrowhead";
+    const shadowId = shape.name.contents + "-shadow";
     // TODO: distinguish between fill opacity and stroke opacity
-
-    const transformStr = svgTransformString(
-      shape.transformation.contents,
-      canvasSize
-    );
-
-    console.log(
-      "path string",
-      toPathString(shape.pathData.contents, canvasSize)
-    );
-
     return (
       <g>
-        <Arrowhead
-          id={leftArrowId}
-          color={strokeColor}
-          opacity={strokeOpacity}
-          style={arrowheadStyle}
-          size={arrowheadSize}
-        />
-        <Arrowhead
-          id={rightArrowId}
-          color={strokeColor}
-          opacity={strokeOpacity}
-          style={arrowheadStyle}
-          size={arrowheadSize}
-        />
+        {shape.leftArrowhead.contents === true ? (
+          <Arrowhead
+            id={leftArrowId}
+            color={strokeColor}
+            opacity={strokeOpacity}
+            style={arrowheadStyle}
+            size={arrowheadSize}
+          />
+        ) : null}
+        {shape.rightArrowhead.contents === true ? (
+          <Arrowhead
+            id={rightArrowId}
+            color={strokeColor}
+            opacity={strokeOpacity}
+            style={arrowheadStyle}
+            size={arrowheadSize}
+          />
+        ) : null}
+        <Shadow id={shadowId} />
         <path
           stroke={strokeColor}
           fill={fillColor}
@@ -115,7 +115,9 @@ class CurveTransform extends React.Component<IGPIProps> {
               ? `url(#${rightArrowId})`
               : ""
           }
-          transform={transformStr}
+          filter={
+            shape.effect.contents === "dropShadow" ? `url(#${shadowId})` : ""
+          }
         >
           <title>{shape.name.contents}</title>
         </path>
@@ -123,4 +125,4 @@ class CurveTransform extends React.Component<IGPIProps> {
     );
   }
 }
-export default CurveTransform;
+export default Curve;
