+++ react-renderer/src/shapes/CurveTransform.tsx
@@ -1,12 +1,12 @@
 import * as React from "react";
-import { toScreen, toHex, Arrowhead, Shadow } from "./Util";
+import { svgTransformString, toHex, Arrowhead } from "utils/Util";
 import { flatten } from "lodash";
-import { IGPIProps } from "./types";
+import { IGPIProps } from "types";
 
 const toCmdString = (cmd: any, canvasSize: [number, number]) => {
   switch (cmd.tag) {
     case "Pt":
-      return "L" + toScreen(cmd.contents, canvasSize).join(" ");
+      return "L" + cmd.contents.join(" ");
     case "CubicBez":
       return pathCommandString("C", cmd.contents, canvasSize);
     case "CubicBezJoin":
@@ -28,25 +28,19 @@ const pathCommandString = (
   command +
   flatten(
     pts.map((coords: [number, number]) => {
-      return toScreen(coords, canvasSize);
+      return coords; // TODO factor this out
     })
   ).join(" ");
 
 const fstCmdString = (pathCmd: any, canvasSize: [number, number]) => {
   if (pathCmd.tag === "Pt") {
-    return "M" + toScreen(pathCmd.contents, canvasSize).join(" ");
+    return "M" + pathCmd.contents.join(" ");
   } else {
     return toCmdString(pathCmd, canvasSize);
   }
 };
 
 const toSubPathString = (commands: any[], canvasSize: [number, number]) => {
-  // TODO: deal with an empty list more gracefully. This next line will crash with undefined head command if empty.
-  if (!commands || !commands.length) {
-    console.error("WARNING: empty path");
-    return "";
-  }
-
   const [headCommand, ...tailCommands] = commands;
   return (
     fstCmdString(headCommand, canvasSize) +
@@ -63,43 +57,49 @@ const toPathString = (pathData: any[], canvasSize: [number, number]) =>
     })
     .join(" ");
 
-class Curve extends React.Component<IGPIProps> {
+class CurveTransform extends React.Component<IGPIProps> {
   public render() {
     const { shape } = this.props;
     const { canvasSize } = this.props;
     const strokeWidth = shape.strokeWidth.contents;
     const strokeColor = toHex(shape.color.contents);
+    console.log("strokeColor", shape.color.contents, strokeColor);
     const fillColor = toHex(shape.fill.contents);
     const strokeOpacity = shape.color.contents.contents[3];
-    const fillOpacity = shape.fill.contents.contents[3];
+    const fillOpacity = shape.fill.contents[3];
     const arrowheadStyle = shape.arrowheadStyle.contents;
     const arrowheadSize = shape.arrowheadSize.contents;
 
     const leftArrowId = shape.name.contents + "-leftArrowhead";
     const rightArrowId = shape.name.contents + "-rightArrowhead";
-    const shadowId = shape.name.contents + "-shadow";
     // TODO: distinguish between fill opacity and stroke opacity
+
+    const transformStr = svgTransformString(
+      shape.transformation.contents,
+      canvasSize
+    );
+
+    console.log(
+      "path string",
+      toPathString(shape.pathData.contents, canvasSize)
+    );
+
     return (
       <g>
-        {shape.leftArrowhead.contents === true ? (
-          <Arrowhead
-            id={leftArrowId}
-            color={strokeColor}
-            opacity={strokeOpacity}
-            style={arrowheadStyle}
-            size={arrowheadSize}
-          />
-        ) : null}
-        {shape.rightArrowhead.contents === true ? (
-          <Arrowhead
-            id={rightArrowId}
-            color={strokeColor}
-            opacity={strokeOpacity}
-            style={arrowheadStyle}
-            size={arrowheadSize}
-          />
-        ) : null}
-        <Shadow id={shadowId} />
+        <Arrowhead
+          id={leftArrowId}
+          color={strokeColor}
+          opacity={strokeOpacity}
+          style={arrowheadStyle}
+          size={arrowheadSize}
+        />
+        <Arrowhead
+          id={rightArrowId}
+          color={strokeColor}
+          opacity={strokeOpacity}
+          style={arrowheadStyle}
+          size={arrowheadSize}
+        />
         <path
           stroke={strokeColor}
           fill={fillColor}
@@ -115,9 +115,7 @@ class Curve extends React.Component<IGPIProps> {
               ? `url(#${rightArrowId})`
               : ""
           }
-          filter={
-            shape.effect.contents === "dropShadow" ? `url(#${shadowId})` : ""
-          }
+          transform={transformStr}
         >
           <title>{shape.name.contents}</title>
         </path>
@@ -125,4 +123,4 @@ class Curve extends React.Component<IGPIProps> {
     );
   }
 }
-export default Curve;
+export default CurveTransform;
