+++ react-renderer/src/shapes/Line.tsx
@@ -1,53 +1,60 @@
 import * as React from "react";
 
-import { IGPIProps } from "./types";
-import { toHex, svgTransformString, Arrowhead } from "./Util";
+import { IGPIProps } from "types";
+import { toScreen, toHex, Arrowhead } from "utils/Util";
 
-class LineTransform extends React.Component<IGPIProps> {
+class Line extends React.Component<IGPIProps> {
   public render() {
     const { shape, canvasSize } = this.props;
     const style = shape.style.contents;
-    const [sx, sy] = [shape.startX.contents, shape.startY.contents];
-    const [ex, ey] = [shape.endX.contents, shape.endY.contents];
+    const [sx, sy] = toScreen(
+      [shape.startX.contents, shape.startY.contents],
+      canvasSize
+    );
+    const [ex, ey] = toScreen(
+      [shape.endX.contents, shape.endY.contents],
+      canvasSize
+    );
+
+    // Rounding for illustrator? Doesn't seem to work
+    /* sx = round2(sx);
+     * sy = round2(sy);
+     * ex = round2(ex);
+     * ey = round2(ey);*/
+
     const path = `M ${sx} ${sy} L ${ex} ${ey}`;
-    const strokeColor = toHex(shape.color.contents);
+    const color = toHex(shape.color.contents);
     const thickness = shape.thickness.contents;
     const strokeDasharray = style === "dashed" ? "7, 5" : "";
-    const strokeOpacity = shape.color.contents.contents[3];
+    const opacity = shape.color.contents.contents[3];
     const arrowheadStyle = shape.arrowheadStyle.contents;
     const arrowheadSize = shape.arrowheadSize.contents;
 
     const leftArrowId = shape.name.contents + "-leftArrowhead";
     const rightArrowId = shape.name.contents + "-rightArrowhead";
-    // TODO: distinguish between fill opacity and stroke opacity
-
-    const transformStr = svgTransformString(
-      shape.transformation.contents,
-      canvasSize
-    );
 
     return (
       <g>
         <Arrowhead
           id={leftArrowId}
-          color={strokeColor}
-          opacity={strokeOpacity}
+          color={color}
+          opacity={opacity}
           style={arrowheadStyle}
           size={arrowheadSize}
         />
         <Arrowhead
           id={rightArrowId}
-          color={strokeColor}
-          opacity={strokeOpacity}
+          color={color}
+          opacity={opacity}
           style={arrowheadStyle}
           size={arrowheadSize}
         />
 
         <path
           d={path}
-          fillOpacity={strokeOpacity}
-          strokeOpacity={strokeOpacity}
-          stroke={strokeColor}
+          fillOpacity={opacity}
+          strokeOpacity={opacity}
+          stroke={color}
           strokeWidth={thickness}
           strokeDasharray={strokeDasharray}
           markerStart={
@@ -58,7 +65,6 @@ class LineTransform extends React.Component<IGPIProps> {
               ? `url(#${rightArrowId})`
               : ""
           }
-          transform={transformStr}
         >
           <title>{shape.name.contents}</title>
         </path>
@@ -66,4 +72,4 @@ class LineTransform extends React.Component<IGPIProps> {
     );
   }
 }
-export default LineTransform;
+export default Line;
