+++ react-renderer/src/shapes/LineTransform.tsx
@@ -1,60 +1,53 @@
 import * as React from "react";
 
-import { IGPIProps } from "./types";
-import { toScreen, toHex, Arrowhead } from "./Util";
+import { IGPIProps } from "types";
+import { toHex, svgTransformString, Arrowhead } from "utils/Util";
 
-class Line extends React.Component<IGPIProps> {
+class LineTransform extends React.Component<IGPIProps> {
   public render() {
     const { shape, canvasSize } = this.props;
     const style = shape.style.contents;
-    const [sx, sy] = toScreen(
-      [shape.startX.contents, shape.startY.contents],
-      canvasSize
-    );
-    const [ex, ey] = toScreen(
-      [shape.endX.contents, shape.endY.contents],
-      canvasSize
-    );
-
-      // Rounding for illustrator? Doesn't seem to work
-      /* sx = round2(sx);
-       * sy = round2(sy);
-       * ex = round2(ex);
-       * ey = round2(ey);*/
-
+    const [sx, sy] = [shape.startX.contents, shape.startY.contents];
+    const [ex, ey] = [shape.endX.contents, shape.endY.contents];
     const path = `M ${sx} ${sy} L ${ex} ${ey}`;
-    const color = toHex(shape.color.contents);
+    const strokeColor = toHex(shape.color.contents);
     const thickness = shape.thickness.contents;
     const strokeDasharray = style === "dashed" ? "7, 5" : "";
-    const opacity = shape.color.contents.contents[3];
+    const strokeOpacity = shape.color.contents.contents[3];
     const arrowheadStyle = shape.arrowheadStyle.contents;
     const arrowheadSize = shape.arrowheadSize.contents;
 
     const leftArrowId = shape.name.contents + "-leftArrowhead";
     const rightArrowId = shape.name.contents + "-rightArrowhead";
+    // TODO: distinguish between fill opacity and stroke opacity
+
+    const transformStr = svgTransformString(
+      shape.transformation.contents,
+      canvasSize
+    );
 
     return (
       <g>
         <Arrowhead
           id={leftArrowId}
-          color={color}
-          opacity={opacity}
+          color={strokeColor}
+          opacity={strokeOpacity}
           style={arrowheadStyle}
           size={arrowheadSize}
         />
         <Arrowhead
           id={rightArrowId}
-          color={color}
-          opacity={opacity}
+          color={strokeColor}
+          opacity={strokeOpacity}
           style={arrowheadStyle}
           size={arrowheadSize}
         />
 
         <path
           d={path}
-          fillOpacity={opacity}
-          strokeOpacity={opacity}
-          stroke={color}
+          fillOpacity={strokeOpacity}
+          strokeOpacity={strokeOpacity}
+          stroke={strokeColor}
           strokeWidth={thickness}
           strokeDasharray={strokeDasharray}
           markerStart={
@@ -65,6 +58,7 @@ class Line extends React.Component<IGPIProps> {
               ? `url(#${rightArrowId})`
               : ""
           }
+          transform={transformStr}
         >
           <title>{shape.name.contents}</title>
         </path>
@@ -72,4 +66,4 @@ class Line extends React.Component<IGPIProps> {
     );
   }
 }
-export default Line;
+export default LineTransform;
