+++ react-renderer/src/shapes/Polygon.tsx
@@ -1,31 +1,28 @@
 import * as React from "react";
-import { toHex, svgTransformString } from "./Util";
-import { IGPIProps } from "./types";
+import { toHex, toPointListString, svgTransformString } from "utils/Util";
+import { IGPIProps } from "types";
 
-class ParallelogramTransform extends React.Component<IGPIProps> {
+class Polygon extends React.Component<IGPIProps> {
   public render() {
     const { shape } = this.props;
     const { canvasSize } = this.props;
 
-    const fillColor = toHex(shape.color.contents);
-    const fillAlpha = shape.color.contents.contents[3];
+    const fillColor = toHex(shape.fillColor.contents);
+    const fillAlpha = shape.fillColor.contents[3];
     const strokeColor = toHex(shape.strokeColor.contents);
-    const strokeAlpha = shape.strokeColor.contents.contents[3];
+    const strokeAlpha = shape.strokeColor.contents[3];
     const thickness = shape.strokeWidth.contents;
+
+    const ptListString = toPointListString(shape.points.contents, canvasSize);
 
     const transformStr = svgTransformString(
       shape.transformation.contents,
       canvasSize
     );
 
-    // The default parallelogram is an axis-aligned unit square centered at the origin
-    // Its position, size, angle, etc. is all set by the Penrose transform
     return (
-      <rect
-        x={-0.5}
-        y={-0.5}
-        width={1.0}
-        height={1.0}
+      <polygon
+        points={ptListString}
         fill={fillColor}
         fillOpacity={fillAlpha}
         stroke={strokeColor}
@@ -35,8 +32,8 @@ class ParallelogramTransform extends React.Component<IGPIProps> {
         transform={transformStr}
       >
         <title>{shape.name.contents}</title>
-      </rect>
+      </polygon>
     );
   }
 }
-export default ParallelogramTransform;
+export default Polygon;
