+++ react-renderer/src/shapes/Rectangle.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import { toScreen, toHex } from "./Util";
-import { IGPIProps } from "./types";
+import { toScreen, toHex } from "utils/Util";
+import { IGPIProps } from "types";
 
 class Rectangle extends React.Component<IGPIProps> {
   public render() {
