+++ react-renderer/src/shapes/SquareTransform.tsx
@@ -1,24 +1,29 @@
 import * as React from "react";
-import { toScreen, toHex } from "./Util";
-import { IGPIProps } from "./types";
+import { svgTransformString, toHex } from "utils/Util";
+import { IGPIProps } from "types";
 
-class Square extends React.Component<IGPIProps> {
+class SquareTransform extends React.Component<IGPIProps> {
   public render() {
     const { shape } = this.props;
     const { canvasSize } = this.props;
-    const [x, y] = toScreen([shape.x.contents, shape.y.contents], canvasSize);
+    // const [x, y] = toScreen([shape.x.contents, shape.y.contents], canvasSize);
     const color = toHex(shape.color.contents);
     const alpha = shape.color.contents.contents[3];
     const strokeColor = toHex(shape.strokeColor.contents);
-    const side = shape.side.contents;
+    const side = 1.0; // shape.side.contents;
     const strokeWidth = shape.strokeWidth.contents;
+    const transformStr = svgTransformString(
+      shape.transformation.contents,
+      canvasSize
+    );
 
     return (
       <rect
-        x={x - side / 2}
-        y={y - side / 2}
+        x={-0.5}
+        y={-0.5}
         width={side}
         height={side}
+        transform={transformStr}
         fill={color}
         fillOpacity={alpha}
         strokeWidth={strokeWidth}
@@ -29,4 +34,4 @@ class Square extends React.Component<IGPIProps> {
     );
   }
 }
-export default Square;
+export default SquareTransform;
