+++ react-renderer/src/utils/Util.tsx
