+++ react-renderer/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": "build/dist",
     "module": "esnext",
     "target": "es5",
-      "lib": ["es5", "es6", "es7", "esnext", "dom"],
+    "lib": ["es5", "es6", "es7", "esnext", "dom"],
     "sourceMap": true,
     "allowJs": true,
     "jsx": "react",
@@ -21,7 +21,10 @@
     "noUnusedLocals": false,
     "skipLibCheck": true, // for Tensorflow JS
     "downlevelIteration": true,
-    "types": ["react", "jest", "node", "tslib"]
+    "types": ["react", "jest", "node", "tslib"],
+    "paths": {
+      "*": ["src/*"]
+    }
   },
   "formatCodeOptions": {
     "baseIndentSize": 0,
